+++ .gitignore
@@ -183,3 +183,5 @@ $RECYCLE.BIN/
 
 # config files
 config/
+
+.idea
